TEST COMMIT

+++ PDStereo/InjeAI/model.py
@@ -544,7 +544,7 @@ def data_generator(dataset, config, shuffle=True, augment=False, augmentation=No
             batch_rpn_bbox[b] = rpn_bbox
             if depth_image_used:
                 batch_images[b] = mold_image(image[...,:3].astype(np.float32), AttrDict({"MEAN_PIXEL": np.array(config.MEAN_PIXEL[:3])}))
-                depth_image = np.resize(image[...,3], ((image.shape[:2]) + (1,)))
+                depth_image = np.resize(image[...,3], ((image.shape[:2]) + (1,))).astype(np.float32)
                 attrDict_MP = AttrDict({"MEAN_PIXEL": config.MEAN_PIXEL[3]})
                 print("DEPTH SHAPE:", depth_image.shape, attrDict_MP.MEAN_PIXEL.shape)
                 batch_depth_images[b] = mold_image(
